pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastquerydr"
version = "0.1.0"
description = "Scaffold for online-efficient dense retrieval experiments"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "faiss-cpu>=1.8",
  "PyYAML>=6.0",
  "psutil>=5.9",
  "torch>=2.2",
  "transformers>=4.39",
  "tqdm>=4.66",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
